LEA (`TestConfigurator`): Fetch VIPER `parameters` on `ViperSource` selection.
